Method argument types update

+++ src/main/java/com/harvestmanager/broadacre/entity/Crop.java
@@ -46,7 +46,7 @@ public class Crop {
         return cropId;
     }
 
-    public void setCropId(int cropId) {
+    public void setCropId(long cropId) {
         this.cropId = cropId;
     }
